made one controller
