+++ src/main/java/aharon/products/ProductsController.java
@@ -12,10 +12,43 @@ public class ProductsController {
     private ProductsService service;
     private JLabel[] imageLabels;
 
-
     public ProductsController(ProductsService service, JLabel[] imageLabels) {
         this.imageLabels = imageLabels;
         this.service = service;
+        listener();
+    }
+
+    private void listener() {
+        for (JLabel label : imageLabels) {
+            label.addMouseListener(new MouseListener() {
+                @Override
+                public void mouseClicked(MouseEvent event) {
+                    JLabel sourceLabel = (JLabel) event.getSource();
+                    Product product = (Product) sourceLabel.getClientProperty("product");
+
+                    if (product != null) {
+                        ProductDetailsFrame detailsFrame = new ProductDetailsFrame(product);
+                        detailsFrame.setVisible(true);
+                    }
+                }
+
+                @Override
+                public void mousePressed(MouseEvent e) {
+                }
+
+                @Override
+                public void mouseReleased(MouseEvent e) {
+                }
+
+                @Override
+                public void mouseEntered(MouseEvent e) {
+                }
+
+                @Override
+                public void mouseExited(MouseEvent e) {
+                }
+            });
+        }
     }
 
     public void display() {
